+++ src/components/Prompt.tsx
@@ -1,5 +1,5 @@
 import PropTypes from "prop-types";
-import React from "react";
+import React, { useEffect } from "react";
 import { useRouterContext } from "../hooks/useRouterContext";
 import { RouterException } from "../RouterException";
 import Lifecycle from "./Lifecycle";
@@ -23,26 +23,20 @@ export default function Prompt({ message, when = true }: PromptProps) {
 	if (!when || context.staticContext) { return null; }
 
 	const method = context.history.block;
+	let release: Function | null = null;
 
-	return (
-		<Lifecycle
-			onMount={(self) => {
-				self.release = method(message);
-			}}
-			onUpdate={(self, prevProps) => {
-				if (prevProps.message !== message && self.release) {
-					self.release();
-					self.release = method(message);
-				}
-			}}
-			onUnmount={(self) => {
-				if (self.release) {
-					self.release();
-				}
-			}}
-			message={message}
-		/>
-	);
+	useEffect(() => {
+		if (release){
+			release();
+		}
+		release = method(message);
+		return () => {
+			if (release){
+				release();
+			}
+		}
+	}, [message])
+	return null;
 }
 
 if (process.env.NODE_ENV !== "production") {